Make i32 the default integer type and fix test suite errors.
These things were closely related.

+++ src/Language/Futhark/Parser.hs
@@ -39,7 +39,7 @@ parseInMonad p file program =
   either (Left . ParseError) Right <$> either (return . Left)
   (evalStateT (evalStateT (runExceptT p) env))
   (alexScanTokens file program)
-  where env = newParserEnv file Int64 Float64
+  where env = newParserEnv file Int32 Float64
 
 parseIncrementalIO :: ParserMonad a -> FilePath -> String
                    -> IO (Either ParseError a)
